requirements.txt
crewai
load_dotenv
crewai_tools